@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2017, Petr Panteleyev <petr@panteleyev.org>
+ * Copyright (c) 2016, 2018, Petr Panteleyev <petr@panteleyev.org>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -25,8 +25,8 @@
  */
 package org.panteleyev.pwdmanager;
 
-import java.util.Objects;
 import javafx.beans.property.SimpleBooleanProperty;
+import java.util.Objects;
 
 public class Category extends Record {
     private SimpleBooleanProperty expanded = new SimpleBooleanProperty();
@@ -43,7 +43,7 @@ public class Category extends Record {
     @Override
     public Category clone() {
         try {
-            Category clone = (Category)super.clone();
+            var clone = (Category) super.clone();
             clone.expanded = new SimpleBooleanProperty(expanded.get());
             return clone;
         } catch (Exception ex) {
@@ -63,9 +63,9 @@ public class Category extends Record {
         }
 
         if (o instanceof Category) {
-            Category that = (Category)o;
+            var that = (Category) o;
             return super.equals(o)
-                && Objects.equals(this.expanded.get(), that.expanded.get());
+                    && Objects.equals(this.expanded.get(), that.expanded.get());
         } else {
             return false;
         }
