@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2017, Petr Panteleyev <petr@panteleyev.org>
+ * Copyright (c) 2016, 2018, Petr Panteleyev <petr@panteleyev.org>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -39,7 +39,7 @@ public enum FieldType {
     private final String name;
 
     FieldType() {
-        ResourceBundle bundle = ResourceBundle.getBundle(BUNDLE);
+        var bundle = ResourceBundle.getBundle(BUNDLE);
         this.name = bundle.getString(name());
     }
 
