@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, Petr Panteleyev <petr@panteleyev.org>
+ * Copyright (c) 2016, 2018, Petr Panteleyev <petr@panteleyev.org>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -25,76 +25,70 @@
  */
 package org.panteleyev.pwdmanager;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.util.ArrayList;
-import java.util.List;
 import javafx.application.Platform;
 import javafx.scene.control.TreeItem;
-import javax.xml.parsers.DocumentBuilder;
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.xml.sax.SAXException;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerConfigurationException;
 import javax.xml.transform.TransformerException;
 import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.dom.DOMSource;
 import javax.xml.transform.stream.StreamResult;
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
-import org.w3c.dom.Node;
-import org.w3c.dom.NodeList;
-import org.xml.sax.SAXException;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.util.ArrayList;
 
 public class Serializer {
     // Attributes
-    private static final String ID_ATTR         = "id";
-    private static final String NAME_ATTR       = "name";
-    private static final String TYPE_ATTR       = "type";
-    private static final String MODIFIED_ATTR   = "modified";
-    private static final String VALUE_ATTR      = "value";
-    private static final String PICTURE_ATTR    = "picture";
-    private static final String EXPANDED_ATTR   = "expanded";
-    private static final String TARGET_ID_ATTR  = "targetId";
+    private static final String ID_ATTR = "id";
+    private static final String NAME_ATTR = "name";
+    private static final String TYPE_ATTR = "type";
+    private static final String MODIFIED_ATTR = "modified";
+    private static final String VALUE_ATTR = "value";
+    private static final String PICTURE_ATTR = "picture";
+    private static final String EXPANDED_ATTR = "expanded";
+    private static final String TARGET_ID_ATTR = "targetId";
 
     // Tags
-    private static final String FIELD           = "field";
-    private static final String FIELDS          = FIELD + "s";
-    private static final String RECORDS         = "records";
+    private static final String FIELD = "field";
+    private static final String FIELDS = FIELD + "s";
+    private static final String RECORDS = "records";
 
     private static final DocumentBuilderFactory DOC_FACTORY;
 
-    public static void serialize(OutputStream out, TreeItem<Record> rootItem) throws IOException, ParserConfigurationException, TransformerConfigurationException, TransformerException {
-		DocumentBuilder docBuilder = DOC_FACTORY.newDocumentBuilder();
+    public static void serialize(OutputStream out, TreeItem<Record> rootItem) throws ParserConfigurationException, TransformerException {
+        var docBuilder = DOC_FACTORY.newDocumentBuilder();
 
-        Document doc = docBuilder.newDocument();
-        Element rootElement = doc.createElement("root");
+        var doc = docBuilder.newDocument();
+        var rootElement = doc.createElement("root");
 
         doc.appendChild(rootElement);
 
-        Element e = serializeTreeItem(doc, rootItem);
+        var e = serializeTreeItem(doc, rootItem);
         rootElement.appendChild(e);
 
-        TransformerFactory transformerFactory = TransformerFactory.newInstance();
-        Transformer transformer = transformerFactory.newTransformer();
-        DOMSource source = new DOMSource(doc);
-        StreamResult result = new StreamResult(out);
+        var transformerFactory = TransformerFactory.newInstance();
+        var transformer = transformerFactory.newTransformer();
+        var source = new DOMSource(doc);
+        var result = new StreamResult(out);
         transformer.transform(source, result);
     }
 
     public static TreeItem<Record> deserialize(InputStream in) throws ParserConfigurationException, SAXException, IOException {
-        DocumentBuilder docBuilder = DOC_FACTORY.newDocumentBuilder();
+        var docBuilder = DOC_FACTORY.newDocumentBuilder();
 
-        Document doc = docBuilder.parse(in);
+        var doc = docBuilder.parse(in);
 
-        Element rootElement = doc.getDocumentElement();
-        NodeList nodes = rootElement.getChildNodes();
+        var rootElement = doc.getDocumentElement();
+        var nodes = rootElement.getChildNodes();
 
         for (int i = 0; i < nodes.getLength(); i++) {
-            Node n = nodes.item(i);
-            if (n instanceof Element && ((Element)n).getTagName().equals("Category")) {
-                return deserializeCategory((Element)n);
+            var n = nodes.item(i);
+            if (n instanceof Element && ((Element) n).getTagName().equals("Category")) {
+                return deserializeCategory((Element) n);
             }
         }
 
@@ -102,9 +96,9 @@ public class Serializer {
     }
 
     public static Element serializeTreeItem(Document doc, TreeItem<Record> treeItem) {
-        Record r = treeItem.getValue();
+        var r = treeItem.getValue();
 
-        Element xmlRecord = doc.createElement(r.getClass().getSimpleName());
+        var xmlRecord = doc.createElement(r.getClass().getSimpleName());
         xmlRecord.setAttribute(ID_ATTR, r.getId());
         xmlRecord.setAttribute(NAME_ATTR, r.getName());
         xmlRecord.setAttribute(TYPE_ATTR, r.getType().name());
@@ -113,43 +107,43 @@ public class Serializer {
 
         // Category
         if (r instanceof Category) {
-            xmlRecord.setAttribute(EXPANDED_ATTR, Boolean.toString(((Category)r).expandedProperty().get()));
+            xmlRecord.setAttribute(EXPANDED_ATTR, Boolean.toString(((Category) r).expandedProperty().get()));
         }
 
         if (r instanceof Link) {
-            xmlRecord.setAttribute(TARGET_ID_ATTR, ((Link)r).getTargetId());
+            xmlRecord.setAttribute(TARGET_ID_ATTR, ((Link) r).getTargetId());
         }
 
         // Card - serialize fields
         if (r instanceof Card) {
-            List<Field> fields = ((Card)r).getFields();
+            var fields = ((Card) r).getFields();
             if (!fields.isEmpty()) {
-                Element fieldsElement = doc.createElement(FIELDS);
+                var fieldsElement = doc.createElement(FIELDS);
                 xmlRecord.appendChild(fieldsElement);
 
                 fields.forEach((f) -> {
-                    Element fe = serializeField(doc, f);
+                    var fe = serializeField(doc, f);
                     fieldsElement.appendChild(fe);
                 });
             }
 
-            String note = ((Card)r).getNote();
-            Element noteElement = doc.createElement("note");
+            var note = ((Card) r).getNote();
+            var noteElement = doc.createElement("note");
             noteElement.setTextContent(note);
             xmlRecord.appendChild(noteElement);
         }
 
         if (r instanceof Note) {
-            xmlRecord.appendChild(doc.createTextNode(((Note)r).getText()));
+            xmlRecord.appendChild(doc.createTextNode(((Note) r).getText()));
         }
 
         // Children if any
         if (!treeItem.getChildren().isEmpty()) {
-            Element children = doc.createElement(RECORDS);
+            var children = doc.createElement(RECORDS);
             xmlRecord.appendChild(children);
 
             treeItem.getChildren().forEach((child) -> {
-                Element e = serializeTreeItem(doc, child);
+                var e = serializeTreeItem(doc, child);
                 children.appendChild(e);
             });
         }
@@ -158,7 +152,7 @@ public class Serializer {
     }
 
     public static Element serializeField(Document doc, Field f) {
-        Element e = doc.createElement(FIELD);
+        var e = doc.createElement(FIELD);
         e.setAttribute(NAME_ATTR, f.getName());
         e.setAttribute(TYPE_ATTR, f.getType().name());
         e.setAttribute(VALUE_ATTR, f.getValue());
@@ -166,77 +160,77 @@ public class Serializer {
     }
 
     public static Field deserializeField(Element e) {
-        String name = e.getAttribute(NAME_ATTR);
-        FieldType type = FieldType.valueOf(e.getAttribute(TYPE_ATTR));
-        String value = e.getAttribute(VALUE_ATTR);
+        var name = e.getAttribute(NAME_ATTR);
+        var type = FieldType.valueOf(e.getAttribute(TYPE_ATTR));
+        var value = e.getAttribute(VALUE_ATTR);
 
         return new Field(type, name, value);
     }
 
     public static TreeItem<Record> deserializeNote(Element element) {
-        String id = element.getAttribute(ID_ATTR);
-        String name = element.getAttribute(NAME_ATTR);
+        var id = element.getAttribute(ID_ATTR);
+        var name = element.getAttribute(NAME_ATTR);
         long modified = Long.valueOf(element.getAttribute(MODIFIED_ATTR));
-        String text = element.getTextContent();
+        var text = element.getTextContent();
 
         return new TreeItem<>(new Note(id, modified, name, text));
     }
 
     public static TreeItem<Record> deserializeLink(Element element) {
-        String targetId = element.getAttribute(TARGET_ID_ATTR);
+        var targetId = element.getAttribute(TARGET_ID_ATTR);
         return new TreeItem<>(new Link(targetId));
     }
 
     public static TreeItem<Record> deserializeCategory(Element element) {
-        String id = element.getAttribute(ID_ATTR);
-        String name = element.getAttribute(NAME_ATTR);
-        RecordType type = RecordType.valueOf(element.getAttribute(TYPE_ATTR));
-        Picture picture = Picture.valueOf(element.getAttribute(PICTURE_ATTR));
+        var id = element.getAttribute(ID_ATTR);
+        var name = element.getAttribute(NAME_ATTR);
+        var type = RecordType.valueOf(element.getAttribute(TYPE_ATTR));
+        var picture = Picture.valueOf(element.getAttribute(PICTURE_ATTR));
         long modified = Long.valueOf(element.getAttribute(MODIFIED_ATTR));
 
-        String expandedString = element.getAttribute(EXPANDED_ATTR);
+        var expandedString = element.getAttribute(EXPANDED_ATTR);
         if (expandedString.isEmpty()) {
             expandedString = "false";
         }
         boolean expanded = Boolean.valueOf(expandedString);
 
-        Category category = new Category(id, modified, name, type, picture, expanded);
-        TreeItem<Record> result = new TreeItem<>(category);
+        var category = new Category(id, modified, name, type, picture, expanded);
+        var result = new TreeItem<Record>(category);
         result.setExpanded(expanded);
 
         category.expandedProperty().bind(result.expandedProperty());
 
-        result.expandedProperty().addListener((x,oldValue,newValue) -> {
+        result.expandedProperty().addListener((x, oldValue, newValue) -> {
             if (oldValue != newValue) {
                 Platform.runLater(() -> MainWindowController.getMainWindow().writeDocument());
             }
         });
 
         // children
-        NodeList childNodes = element.getChildNodes();
+        var childNodes = element.getChildNodes();
         for (int i = 0; i < childNodes.getLength(); i++) {
-            Node n = childNodes.item(i);
+            var n = childNodes.item(i);
 
-            if (n instanceof Element && ((Element)n).getTagName().equals(RECORDS)) {
-                NodeList records = n.getChildNodes();
+            if (n instanceof Element && ((Element) n).getTagName().equals(RECORDS)) {
+                var records = n.getChildNodes();
                 for (int j = 0; j < records.getLength(); j++) {
-                    Node r = records.item(j);
+                    var r = records.item(j);
 
                     if (r instanceof Element) {
-                        Element re = (Element)r;
+                        var re = (Element) r;
 
                         TreeItem<Record> recordItem = null;
                         switch (re.getTagName()) {
-                            case "Category" :
+                            case "Category":
                                 recordItem = deserializeCategory(re);
                                 break;
-                            case "Card" :
+                            case "Card":
                                 recordItem = deserializeCard(re);
                                 break;
-                            case "Note" :
+                            case "Note":
                                 recordItem = deserializeNote(re);
                                 break;
-                            case "Link" :
+                            case "Link":
                                 recordItem = deserializeLink(re);
                                 break;
                         }
@@ -257,24 +251,24 @@ public class Serializer {
     }
 
     public static TreeItem<Record> deserializeCard(Element element) {
-        String id = element.getAttribute(ID_ATTR);
-        String name = element.getAttribute(NAME_ATTR);
-        Picture picture = Picture.valueOf(element.getAttribute(PICTURE_ATTR));
+        var id = element.getAttribute(ID_ATTR);
+        var name = element.getAttribute(NAME_ATTR);
+        var picture = Picture.valueOf(element.getAttribute(PICTURE_ATTR));
         long modified = Long.valueOf(element.getAttribute(MODIFIED_ATTR));
 
         // fields
-        NodeList fList = element.getElementsByTagName(FIELD);
-        List<Field> fields = new ArrayList<>(fList.getLength());
+        var fList = element.getElementsByTagName(FIELD);
+        var fields = new ArrayList<Field>(fList.getLength());
         for (int i = 0; i < fList.getLength(); i++) {
-            Field f = deserializeField((Element)fList.item(i));
+            var f = deserializeField((Element) fList.item(i));
             fields.add(f);
         }
 
         // note
-        String note = "";
-        NodeList notes = element.getElementsByTagName("note");
+        var note = "";
+        var notes = element.getElementsByTagName("note");
         if (notes.getLength() > 0) {
-            Element noteElement = (Element)notes.item(0);
+            var noteElement = (Element) notes.item(0);
             note = noteElement.getTextContent();
         }
 
