@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2017, Petr Panteleyev <petr@panteleyev.org>
+ * Copyright (c) 2016, 2018, Petr Panteleyev <petr@panteleyev.org>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -25,11 +25,7 @@
  */
 package org.panteleyev.pwdmanager.test;
 
-import java.util.Arrays;
-import java.util.UUID;
 import javafx.scene.control.TreeItem;
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
 import org.panteleyev.pwdmanager.Card;
 import org.panteleyev.pwdmanager.Category;
 import org.panteleyev.pwdmanager.Field;
@@ -37,15 +33,16 @@ import org.panteleyev.pwdmanager.FieldType;
 import org.panteleyev.pwdmanager.Link;
 import org.panteleyev.pwdmanager.Note;
 import org.panteleyev.pwdmanager.Picture;
-import org.panteleyev.pwdmanager.Record;
 import org.panteleyev.pwdmanager.RecordType;
 import org.panteleyev.pwdmanager.Serializer;
 import org.testng.Assert;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import java.util.Arrays;
+import java.util.UUID;
 
 public class TestSerializer {
     private DocumentBuilderFactory docFactory;
@@ -59,69 +56,69 @@ public class TestSerializer {
 
     @Test
     public void testCardSerialization() throws Exception {
-        Card card = new Card(UUID.randomUUID().toString(), Picture.AMEX, Arrays.asList(
-            new Field(FieldType.STRING, UUID.randomUUID().toString(), UUID.randomUUID().toString()),
-            new Field(FieldType.HIDDEN, UUID.randomUUID().toString(), UUID.randomUUID().toString())
+        var card = new Card(UUID.randomUUID().toString(), Picture.AMEX, Arrays.asList(
+                new Field(FieldType.STRING, UUID.randomUUID().toString(), UUID.randomUUID().toString()),
+                new Field(FieldType.HIDDEN, UUID.randomUUID().toString(), UUID.randomUUID().toString())
         ), UUID.randomUUID().toString());
 
-        Document doc = docBuilder.newDocument();
+        var doc = docBuilder.newDocument();
 
-        Element e = Serializer.serializeTreeItem(doc, new TreeItem<>(card));
+        var e = Serializer.serializeTreeItem(doc, new TreeItem<>(card));
 
-        TreeItem<Record> restored = Serializer.deserializeCard(e);
+        var restored = Serializer.deserializeCard(e);
 
         Assert.assertEquals(restored.getValue(), card);
     }
 
     @Test
     public void testCategorySerialization() throws Exception {
-        Category category = new Category(UUID.randomUUID().toString(), RecordType.EMPTY, Picture.FOLDER);
+        var category = new Category(UUID.randomUUID().toString(), RecordType.EMPTY, Picture.FOLDER);
 
-        Document doc = docBuilder.newDocument();
+        var doc = docBuilder.newDocument();
 
-        Element e = Serializer.serializeTreeItem(doc, new TreeItem<>(category));
-        TreeItem<Record> restored = Serializer.deserializeCategory(e);
+        var e = Serializer.serializeTreeItem(doc, new TreeItem<>(category));
+        var restored = Serializer.deserializeCategory(e);
         Assert.assertEquals(restored.getValue(), category);
     }
 
-    @DataProvider(name="testFieldToFromJson")
+    @DataProvider(name = "testFieldToFromJson")
     public Object[][] testFieldToFromJsonDataProvider() {
-        return new Object[][] {
-            { new Field(FieldType.STRING, UUID.randomUUID().toString(), UUID.randomUUID().toString()) },
-            { new Field(FieldType.HIDDEN, UUID.randomUUID().toString(), UUID.randomUUID().toString()) },
-            { new Field(FieldType.LINK, UUID.randomUUID().toString(), "1024") },
+        return new Object[][]{
+                {new Field(FieldType.STRING, UUID.randomUUID().toString(), UUID.randomUUID().toString())},
+                {new Field(FieldType.HIDDEN, UUID.randomUUID().toString(), UUID.randomUUID().toString())},
+                {new Field(FieldType.LINK, UUID.randomUUID().toString(), "1024")},
         };
     }
 
-    @Test(dataProvider="testFieldToFromJson")
+    @Test(dataProvider = "testFieldToFromJson")
     public void testFieldSerialization(Field field) throws Exception {
-        Document doc = docBuilder.newDocument();
+        var doc = docBuilder.newDocument();
 
-        Element e = Serializer.serializeField(doc, field);
-        Field restored = Serializer.deserializeField(e);
+        var e = Serializer.serializeField(doc, field);
+        var restored = Serializer.deserializeField(e);
 
         Assert.assertEquals(restored, field);
     }
 
     @Test
     public void testNoteSerialization() throws Exception {
-        Note note = new Note(UUID.randomUUID().toString(), UUID.randomUUID().toString());
+        var note = new Note(UUID.randomUUID().toString(), UUID.randomUUID().toString());
 
-        Document doc = docBuilder.newDocument();
+        var doc = docBuilder.newDocument();
 
-        Element e = Serializer.serializeTreeItem(doc, new TreeItem<>(note));
-        TreeItem<Record> restored = Serializer.deserializeNote(e);
+        var e = Serializer.serializeTreeItem(doc, new TreeItem<>(note));
+        var restored = Serializer.deserializeNote(e);
         Assert.assertEquals(restored.getValue(), note);
     }
 
     @Test
     public void testLinkSerialization() throws Exception {
-        Link link = new Link(UUID.randomUUID().toString());
+        var link = new Link(UUID.randomUUID().toString());
 
-        Document doc = docBuilder.newDocument();
+        var doc = docBuilder.newDocument();
 
-        Element e = Serializer.serializeTreeItem(doc, new TreeItem<>(link));
-        TreeItem<Record> restored = Serializer.deserializeLink(e);
+        var e = Serializer.serializeTreeItem(doc, new TreeItem<>(link));
+        var restored = Serializer.deserializeLink(e);
         Assert.assertEquals(restored.getValue(), link);
     }
 }
